fix: connect to real TXTDiary/TXTBrain/TXTAIMemory (schema drift + wrong ports)
3개 소스 앱과의 연결이 "된 것처럼 보이지만 실제로 안 되는" 문제의 근본 원인을
실제 서비스를 상대로 조사해서 고쳤다.

발견한 것:
- 기본 포트가 전부 추측값(4001/4002/4003)이었다. 실측: TXTBrain=8811,
  TXTAIMemory keyword-api=47531, TXTDiary=47821(미실행), TXTSpace-hub=47540.
- 실배포 API 응답이 PRD 원안 스키마와 갈라져 있었다: text→keyword,
  cooccurrence→cooccurrences, frequency/count가 부동소수, source가 봉투가
  아니라 키워드 항목마다 붙음(허브 응답), normalized_text 일부 누락.
- sync_source가 파싱 실패를 조용히 삼켜서(Err(_) => {}) "연결은 됐는데
  아무 일도 안 일어나는" 것처럼 보이는 버그가 있었다.
- TXTSpace-hub(47540)가 이미 3소스를 토큰 없이 통합 재노출하고 있었다 —
  X1이 계획과 다른 모양(허브 애그리게이션)으로 먼저 와 있었던 셈.

고친 것:
- core/models.rs: Keyword에 serde alias(keyword/cooccurrences)로 실배포
  스키마 수용, frequency/count를 f64로, 항목별 source 필드 추가,
  normalized_text 누락 시 text로 백필.
- core/source.rs: merge_keywords가 항목별 source를 우선.
- app/pipeline.rs: 파싱 실패를 실제 에러 메시지로 표면화(더 이상 안 삼킴).
- app: 소스 선택에 TXTSpace-hub(4번째 옵션, 토큰 불필요) 추가, 실측 포트로
  기본값 교정, 5개 언어 라벨/힌트 추가.

검증: 실제 hub를 상대로 한 라이브 통합 테스트(#[ignore], cargo test --ignored)
로 190개 키워드·3소스 정상 태깅 확인. 실제 앱 데이터 DB에도 반영해 195개
키워드로 즉시 사용 가능한 상태로 만듦. core 36/36·app 1+1ignored 테스트 통과.

VERSION 0.2.0 → 0.2.1 (버그픽스). 통합 스펙 §6에 실측 관찰 기록.

+++ app/src-tauri/src/dto.rs
@@ -35,7 +35,7 @@ pub struct KeywordDto {
     pub source_label: String,
     pub text: String,
     pub normalized_text: String,
-    pub frequency: u64,
+    pub frequency: f64,
     pub avg_emotion_score: f32,
     pub deeplink: String,
 }
